Use globals for auth_uri lookups
public_auth_uri, internal_auth_uri, and admin_identity_uri are used in
many places, but are each looked up and calculated on their own. This makes
it difficult to ensure consistency, or over ride it if desired.

Here we move from the individual calculations and instead propogate the
paramaters into hiera so that we can use the pre-created values. At the same
time, we look to see if we already received this structure in the event
that they where overridden.

Closes-bug: #1567143
Signed-off-by: Andrew Woodward <awoodward@mirantis.com>
Change-Id: I90d8d030d345b628432fc6ee8de19a656ea532ea