Update stdlib to 4.12.0
The upstream puppet-ceph module is leveraging the service_provider fact
which was added to stdlib as part of 4.10.0. This is causing the wrong
service start method to be used for radosgw.  This change updates stdlib
to pull in the service_provider fact so the correct service management
method is used.

Change-Id: Ib81b800504baa968a08278e5adc3fdf6fb0975c0
Closes-Bug: 1580656